requirements.txt
fastavro
requests
pandas
pytest
google-cloud-core
google-cloud-pubsub
google-cloud-bigquery
pymongo
elasticsearch
sphinx
